+++ changelog.md
@@ -1,9 +1,18 @@
 Version History
 ===============
 
-1.1
----
-Initial release
+Version 1
+---------
+
+### 1.1.2
+
+11 Jan 2018
+
+* Adds calibration that's a linear sweeps as one segment rather than one point per segment. This can cause slow behavior and/or timeouts for calibrations with many points.
 
 ### 1.1.1
-Fixed "*Frequency range error" message for valid calibrations with crossover frequency - crossover frequency now correctly detected and applied in GUI.
+* "\*Frequency range error" for valid calibrations and crossover frequencies fixed.
+
+### 1.1
+
+Initial release